CAMEL-13444 Renaming package
